Deal with sample data more realistically
Even though the sample image and video data is included as code in order
to avoid fetching it from a separate server, it should still be treated
as if it were external data with a potentially unknown structure. These
changes take us part of the way there, by actually parsing strings as
JSON, and typing the parsed data as `unknown` instead of `any`. That
way, we're forced to explicity coerce the data to the proper type when
it's used. (A further improvement would be to actually verify the data
structure instead of just blindly applying the type coercion, of
course.)