Migrate to standalone components
Used the `@angular/core:standalone` schematic (available as of Angular
v15.2.0) as described in the standalone components migration guide
<https://angular.io/guide/standalone-migration> to migrate the
application to standalone components. Everything worked on the first
try, believe it or not.

Once the automatic migration was complete, I did some light reformatting
of the updated code, and fixed a linter complaint, but no important
changes to the code were necessary. Some additional code formatting
would still be good, but I'll deal with that later.